+++ kernel/nf4_grouped.py
@@ -738,7 +738,10 @@ def _gemm_nf4_grouped(
     pid_m = tl.program_id(0)
     pid_n = tl.program_id(1)
 
-    row0 = tl.load(t_row0_ptr + pid_m)
+    # int64 row base: (row0 + offs_m) * K / * N index the activation and
+    # output rows and wrap once T * max(K, N) reaches 2^31 elements; the cast
+    # is one scalar per program (docs/KERNEL_CONTRACT.md, Boundaries).
+    row0 = tl.load(t_row0_ptr + pid_m).to(tl.int64)
     rows = tl.load(t_rows_ptr + pid_m)
     # The static-budget tile table (build_group_tiles_device) pads to
     # ceil(R/BM) + E tiles and marks the padding rows == 0. The store
@@ -1471,7 +1474,9 @@ def _dgrad_nf4_grouped(
     pid_m = tl.program_id(0)
     pid_k = tl.program_id(1)
 
-    row0 = tl.load(t_row0_ptr + pid_m)
+    # int64 row base, as in the forward: grad_out / grad_a row offsets wrap
+    # once T * max(K, N) reaches 2^31 elements.
+    row0 = tl.load(t_row0_ptr + pid_m).to(tl.int64)
     rows = tl.load(t_rows_ptr + pid_m)
     grp = tl.load(t_group_ptr + pid_m)
     eid = tl.load(expert_ids_ptr + grp)
